pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.15.0",
    "fastapi-users>=14.0.1",
    "fastapi>=0.115.5",
    "python-dotenv>=1.0.1",
    "scalar-fastapi>=1.0.3",
    "sqlalchemy>=2.0.37",
    "uvicorn>=0.32.1",
    "requests>=2.32.3",
    "redis>=5.2.1",
    "pillow>=11.1.0",
    "asyncpg>=0.30.0",
    "websockets>=14.2",
    "jinja2>=3.1.5",
    "aioredis>=2.0.1",
    "setuptools>=75.8.0",
]
